Faz flush

+++ src/resources/StreamHdfs.js
@@ -16,7 +16,7 @@ class StreamHdfs extends StreamResource {
     return `hdfs dfs -put -f ${osFilePath} ${hdfsPath}`
   }
 
-  async insertData ({ data, outName, bucket, append }) {
+  async insertData ({ data, outName, bucket, append, flush }) {
     const osPath = `/tmp/${bucket.name}_${outName}`
 
     // file exists
@@ -31,16 +31,21 @@ class StreamHdfs extends StreamResource {
       : FileSystem.write(outData, osPath)
 
     await promise
-    // certificando que o path no hdfs está criado
-    await OS.run(this.createHdfsFilePath(bucket.name))
-    await OS.run(this.shellCommand(bucket.name, outName, osPath))
-      .catch(async err => {
+
+    // Salva no hdfs quando não estiver fazendo mais append
+    if (flush) {
+      // certificando que o path no hdfs está criado
+      await OS.run(this.createHdfsFilePath(bucket.name))
+      await OS.run(this.shellCommand(bucket.name, outName, osPath)).catch(async err => {
         await FileSystem.delete(osPath)
         throw err
       })
+    }
 
     // remove o arquivo temporário do so
-    return FileSystem.delete(osPath)
+    return flush
+      ? FileSystem.delete(osPath)
+      : Promise.resolve()
   }
 }
